change how the datacenter is found
Instead of using serviceInstance.find_datacenter to find the datacenter by name, loop through rootFolder.children looking for a child that has that name and is a datacenter.
The reason to do this is that rootFolder.children is populated with the children that the user has permissions to see. serviceInstance.find_datacenter does rootFolder.traverse which requires the permission to search through all the children