Improve the way we find a host
Submitted by Learath (Nicholas Brisebois) from Dell SecureWorks. See
https://github.com/ezrapagel/knife-vsphere/issues/162

This patch improves the way hosts are located when creating the original
clone spec. Before this patch we'd simply grab the first host we saw
even if it was in maintenance mode or was missing resources needed by
the VM. Now, do some more stringent checking such as removing hosts in
maintenance mode, and check to make sure the user isn't asking for a
datastore that has no available hosts.