Let get_config accept a value of false
get_config checks in both the config hash and the Chef::Config hash for
the given option. Why? I'm not sure, but it's there. It uses the
traditional `val = a || b`, so that if a is nil then the value of b is
used. But if a is intentionally false, it still chooses b because both
are falsey.