clone: Fix waiting for address conditional
Currently the wait condition for the availability of IP addresses is an
and (&&) condition on both vm.guest.net.empty? and !vm.guest.ipAddress,
however this can lead to a race condition where rbvmomi will see the IP
address returned by vsphere before it has the full NIC information,
leading to a NilClass issue in later parts of the logic. I'd imagine
that the intention here was to wait for both to pass, so this needs to
be changed to an or (||).