Address mac_list problem.
Clean up a bit -- if we're going to treat auto and blank as the same for
the mac address then make it the default and get rid of one case. Also
take this opportunity to refactor a tiny bit so that the guard clause
does an early abort and then we set up the data rather than it being
wrapped in an else.

Fixes #277