Fix bug with no cmacs or cips; tests; refactor
I was sent a bug privately where cloning a VM with no --cmacs or
--cips triggered an abort. When I updated --cmacs to default to `auto` I didn't change that guard clause, so the code always assumed there was a cmac specified. So let's get away from checking strings ;)

Also took this opportunity to dive into the mocking necessary to test a clone. There's a lot of mocking the SUT but much of it is unavoidable. I moved some of the code around so that the mocks were working on smaller pieces of code, and I hope that I'll be able to test those pieces of code more thoroughly.

These tests in the cloning code do set us up to be able to test the clone specs being sent with vmware.