Default to vmname in spec; cleanup
Longstanding thing where the code would default to vmname for the
customization hostname, but the previous conditional would never let it
get there. I didn't see when it was changed but it was caused by the
conditional looking for a `nil` `identity` but the `identity` was set to
a blank `CustomizationIdentitySettings`.

Wrote some tests around that and used those validations to clean up the
code slightly, along with some breadcrumbs for future TODOs.