Refactor (#278)
* Fix bug with no cmacs or cips; tests; refactor

I was sent a bug privately where cloning a VM with no --cmacs or
--cips triggered an abort. When I updated --cmacs to default to `auto` I didn't change that guard clause, so the code always assumed there was a cmac specified. So let's get away from checking strings ;)

Also took this opportunity to dive into the mocking necessary to test a clone. There's a lot of mocking the SUT but much of it is unavoidable. I moved some of the code around so that the mocks were working on smaller pieces of code, and I hope that I'll be able to test those pieces of code more thoroughly.

These tests in the cloning code do set us up to be able to test the clone specs being sent with vmware.

* Refactor LocationSpec creation and fix linked-clone

Issue #279 showed some silliness where we overwrite a perfectly good
location spec and end up with something that doesn't work. #280 fixes
one, this does the other. Also hoist the object creation to the top and
clean up the assignment of the pool.

* Finally able to test the spec contents

* Default to vmname in spec; cleanup

Longstanding thing where the code would default to vmname for the
customization hostname, but the previous conditional would never let it
get there. I didn't see when it was changed but it was caused by the
conditional looking for a `nil` `identity` but the `identity` was set to
a blank `CustomizationIdentitySettings`.

Wrote some tests around that and used those validations to clean up the
code slightly, along with some breadcrumbs for future TODOs.

* Send nil spec when disabling customizations and no cspec is present

However we still need to send the server side cspec in this case, which
complicates the logic here. So:

customizations disabled and no cspec - customization = nil
customizations disabled and a cspec -  customization = cspec

Fixes #287