Update vm find to support the formatter
The output of this command has been hard to parse and brittle as we add
new things. Convert this to use the output formatter so that you can
request JSON and then get your output that way.

Also deprecates --ips which was remarkably close to --ip and really was
more about the networks than the IPs.