Proceed to bootstrap if node exists. (#417)
* Proceed to bootstrap if the node already exists

Fixes #416. If the node exists in vsphere, proceed to bootstrap rather
than bail out.

Chef will still prompt the user to overwrite the node and client, so if
for some reason this wasn't intentional, there's a safe way to bail.

I'd considered making this an option but thinking to the way I use this
command it's usually going to be the desired approach. In fact I needed
it twice this morning, leading to me making the change.

* Rubocop