Move vsphere vm find to use container view
Use the container view api in vsphere to retrieve only the info we need
about the VMs so we can search more quickly.

This introduces a breaking change in that the folder is no longer
default. You need to use --short-path to get the folder name.

Update Docs

Safe navigation operator is only in Ruby 2.3.0+. I know it's old but 2.2
is still in heavy use and this doesn't make the code much more readable.

Refactor full_path_to

Improve some testing